Visuo-tactile contrastive learning/requirements.txt
appdirs==1.4.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
debugpy==1.6.7.post1
docker-pycreds==0.4.0
gitdb==4.0.10
GitPython==3.1.34
idna==3.4
imageio==2.31.2
lazy-loader==0.3
networkx==3.1
numpy==1.24.4
opencv-python==4.8.0.76
packaging==23.1
pathtools==0.1.2
Pillow==10.0.0
protobuf==4.24.2
psutil==5.9.5
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
sentry-sdk==1.30.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
tifffile==2023.7.10
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
typing-extensions==4.7.1
urllib3==2.0.4
wandb==0.15.9
